feature[ats_VVMH0]: Upgrade to use Tycho 0.14.1

+++ plugins/org.eclipse.osee.framework.core.model/src/org/eclipse/osee/framework/core/model/IOseeStorable.java
@@ -18,7 +18,7 @@ import org.eclipse.osee.framework.core.exception.OseeCoreException;
  */
 public interface IOseeStorable {
 
-   public final static Integer UNPERSISTED_VALUE = Short.MIN_VALUE;
+   public final static Integer UNPERSISTED_VALUE = (int) Short.MIN_VALUE;
 
    int getId();
